requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
pyyaml>=6.0.1
requests
django-cors-headers
